+++ terminice_core/lib/src/src.dart
@@ -2,3 +2,7 @@ export 'io/key_bindings.dart';
 export 'io/key_events.dart';
 export 'io/terminal_control.dart';
 export 'io/terminal_info.dart';
+export 'navigation/focus_navigator.dart';
+export 'navigation/grid_navigator.dart';
+export 'navigation/list_navigator.dart';
+export 'navigation/selection_controller.dart';